test/unit/modules/voting/commands/create_vote_command.spec.ts
import { CreateVoteCommand } from '../../../../../src/app/modules/voting/commands/create_vote_command';

describe('CreateVoteCommand', () => {
  let command: CreateVoteCommand;

	beforeEach(() => {
		command = new CreateVoteCommand();
	});

	describe('constructor', () => {
		it('should have valid name', () => {
			expect(command.name).toEqual('createVote');
		});

		it('should have valid schema', () => {
			expect(command.schema).toMatchSnapshot();
		});
	});

	describe('verify', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('execute', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
